Implement client UI and interaction with client and server and database.

+++ src/network/client/client.h
@@ -4,4 +4,4 @@
 
 void cli_main();
 
-bool cli_req_id_by_name(const char* name, u16* buf);
+bool cli_req_id_by_name(const char* name, i32* buf);
